+++ opencl/opencl.h
@@ -9,38 +9,38 @@ typedef struct	s_jul
 {
 	int		x_dim;
 	int		y_dim;
-	double	zx;
-	double	zy;
-	double	ci;
-	double	cr;
-	double	x_origin;
-	double	y_origin;
+	float	zx;
+	float	zy;
+	float	ci;
+	float	cr;
+	float	x_origin;
+	float	y_origin;
 	size_t	width;
 	size_t	height;
 }				t_jul;
 
 typedef struct	s_krn
 {
-	double	xmin;
-	double	xmax;
-	double	ymin;
-	double	ymax;
-	double	xoffset;
-	double	yoffset;
+	float	xmin;
+	float	xmax;
+	float	ymin;
+	float	ymax;
+	float	xoffset;
+	float	yoffset;
 	int		iter;
-	double	red;
-	double	green;
-	double	blue;
+	float	red;
+	float	green;
+	float	blue;
 }				t_krn;
 
 typedef struct	s_mdl
 {
-	double	x;
-	double	y;
+	float	x;
+	float	y;
 	int		x_dim;
 	int		y_dim;
-	double	x_origin;
-	double	y_origin;
+	float	x_origin;
+	float	y_origin;
 	size_t	width;
 	size_t	height;
 }				t_mdl;
